backend/requirements.txt
fastapi==0.135.3
uvicorn[standard]==0.43.0
pydantic==2.12.5
pydantic-settings==2.13.1
pyjwt[crypto]==2.10.1
bcrypt==4.2.1
python-multipart==0.0.22
python-dotenv==1.0.1
textstat==0.7.13

# RAG pipeline
chromadb==1.5.5
sentence-transformers>=2.7.0
beautifulsoup4>=4.12.0
requests>=2.31.0

# LLM providers
openai==2.30.0